Reduce surface area under ReportResultStage claim
Hangs have been observed when Executors could not claim
ReportResultStage, but ReportResultStage was waiting to take results.
Outside of a possible notify failure, something that could cause this,
leaving the ReportResultStage claimed by noone, would be an exception
during the claimed period, until the queue insertion. This change
reduces the amount of work performed under that context, and simply
claim->put for successful claim and completed operation. The
InterruptedException listed in documentation for the put is guarded
against with a release as well.