Clarify write cancellation and close behaviors.
A successful close on a putImpl output stream implies a reference to the
file has been obtained.
Any unsuccessful close, including an incomplete blob entry that does not
delete the writeFile, will throw an exception, likely an IOException,
and no reference to the entry should be assumed.
A Write reset will issue a cancel on the putImpl output stream if
available, and will delete any write file in progress.
Writes and writesInProgress will expire based on access with a 1 hour
time limit. The Write will expire through reset.