Introduce worker CAS delegation
Worker CAS specification is now a sequence of CASConfig descriptors.
These will be applied in sequence, with each CAS delegating to the one
after it (in order).

Delegation mandates that:
 * findMissingBlobs/contains defer to a delegate for any still-missing blobs
 * newInput/get defer to the delegate when missing and read-through into
 the current CAS
 * Write/put push evicted content into delegates before unregistering
 it from the current CAS

Delegates are also be considered first before the remote mechanisms
when fetching content for path population, used by input fetches (and
*not* by CAS/BS presentation).